split auth tests
